+++ CHANGELOG.md
@@ -5,7 +5,10 @@ Version 2.0.9 (coming soon...)
 -------
 
 * Added pinch and tap zooming
-* Use scalable SVG image as a location marker. User can change the color of the marker from preferences.
+* Now uses scalable SVG image as a location marker.
+* User can change the color and the size of the map marker and
+municipality points from preferences.
+* Shows the name of the municipality when long pressing on a municipality point
 * Major code refactorings. Now uses RoboGuice 2 for dependency injection
 and Robotium for automatic black-box testing.